+++ custom_components/pollenlevels/const.py
@@ -5,6 +5,7 @@ CONF_API_KEY = "api_key"
 CONF_LATITUDE = "latitude"
 CONF_LONGITUDE = "longitude"
 CONF_UPDATE_INTERVAL = "update_interval"
+CONF_LANGUAGE_CODE = "language_code"
 
 # Default update interval in hours
 DEFAULT_UPDATE_INTERVAL = 6